Add genetic code feature to use custom translation tables
