+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "atglib"
-version = "0.1.4"
+version = "0.2.0"
 edition = "2021"
 authors = ["Jonas Marcello <jonas.marcello@esbme.com>"]
 description = "A library to handle transcripts for genomics and transcriptomics"
