+++ src/models/mod.rs
@@ -1,6 +1,7 @@
 //! Holds the main data models and structs that are used in ATG
 //!
 
+mod aminoacid;
 mod codon;
 mod exon;
 mod frame;
@@ -13,23 +14,13 @@ mod utils;
 pub use crate::models::codon::Codon;
 pub use crate::models::exon::Exon;
 pub use crate::models::frame::Frame;
+pub use crate::models::aminoacid::AminoAcid;
 pub use crate::models::genetic_code::GeneticCode;
 pub use crate::models::sequence::{Nucleotide, Sequence};
 pub use crate::models::transcript::{Coordinate, CoordinateVector, Transcript, TranscriptBuilder};
 pub use crate::models::transcripts::Transcripts;
 pub use crate::models::utils::{CdsStat, Strand, TranscriptRead, TranscriptWrite};
 
-const fn start_codons() -> [sequence::Nucleotide; 3] {
-    [Nucleotide::A, Nucleotide::T, Nucleotide::G]
-}
-
-pub const START_CODON: [sequence::Nucleotide; 3] = start_codons();
-
-pub const STOP_CODONS: [[sequence::Nucleotide; 3]; 3] = [
-    [Nucleotide::T, Nucleotide::A, Nucleotide::G],
-    [Nucleotide::T, Nucleotide::A, Nucleotide::A],
-    [Nucleotide::T, Nucleotide::G, Nucleotide::A],
-];
 
 #[cfg(test)]
 mod tests {
